CMakeLists.txt: copy resources directory to test directory
the xml files inside the resources are to be used for the
run of xmltest
they were not getting copied from the source directory to the directory
where xmltest was installed

fixes the below error when code built with cmake
./xmltest
Error opening test file 'dream.xml'.
Is your working directory the same as where
the xmltest.cpp and dream.xml file are?

Signed-off-by: Nisha Parrakat <nisha.m.parrakat@bmw.de>
Signed-off-by: Nisha Parrakat <nishaparrakat@gmail.com>